retorno Lista

+++ src/br/com/caelum/projetocdc/ComparadorUltimoLancamento.java
@@ -6,8 +6,8 @@ public class ComparadorUltimoLancamento implements Comparator<Livro> {
 
 	@Override
 	public int compare(Livro livroUm, Livro livroDois) {
-		int valor = (int) livroDois.getDataLancamento().getTimeInMillis() - (int) livroUm.getDataLancamento().getTimeInMillis();
-		return valor;
+		long valor = livroDois.getDataLancamento().getTimeInMillis() - livroUm.getDataLancamento().getTimeInMillis();
+		return (int)valor;
 	}
 
 }
